requirement.txt
# Core
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
imbalanced-learn==0.12.0
xgboost==2.0.3
shap==0.45.1
seaborn==0.13.2
matplotlib==3.8.3
tqdm==4.66.2
joblib==1.3.2

# Audio processing
librosa==0.10.1
soundfile==0.12.1
pyworld==0.3.2

# NLP
spacy==3.7.4
https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl

# Alignment
textgrid==1.5
openai-whisper
# MFA (install separately)
# conda create -n aligner -c conda-forge montreal-forced-aligner

# Other
ipython==8.22.2